Add `fuel-data` as dependency
* Remove `Storage` and `MerkleStorage` implementations
* Use `Cow` for storage read operations
* Implement default trait methods for common storage operations
