+++ .github/workflows/cargo_test.yml
@@ -25,6 +25,7 @@ jobs:
       with:
         ssh-private-key: |
           ${{ secrets.PRIVATE_KEY_FUEL_ASM }}
+          ${{ secrets.PRIVATE_KEY_FUEL_DATA }}
           ${{ secrets.PRIVATE_KEY_FUEL_TX }}
 
     - name: Install toolchain
@@ -54,6 +55,12 @@ jobs:
       with:
         command: test
         args: --verbose
+
+    - name: Run tests random
+      uses: actions-rs/cargo@v1
+      with:
+        command: test
+        args: --verbose --features random
 
     - name: Run debug tests
       uses: actions-rs/cargo@v1
