+++ Cargo.toml
@@ -10,6 +10,7 @@ description = "FuelVM interpreter."
 
 [dependencies]
 fuel-asm = { git = "ssh://git@github.com/FuelLabs/fuel-asm.git" }
+fuel-data = { git = "ssh://git@github.com/FuelLabs/fuel-data.git" }
 fuel-tx = { git = "ssh://git@github.com/FuelLabs/fuel-tx.git" }
 itertools = "0.10"
 secp256k1 = { version = "0.20", features = ["recovery"] }
@@ -22,4 +23,30 @@ rand = "0.8"
 
 [features]
 debug = []
-serde-types = ["serde", "fuel-asm/serde-types", "fuel-tx/serde-types"]
+random = [ "fuel-data/random", "fuel-tx/random" ]
+serde-types = [ "fuel-asm/serde-types", "fuel-data/serde-types", "fuel-tx/serde-types", "serde" ]
+
+[[test]]
+name = "test-blockchain"
+path = "tests/blockchain.rs"
+required-features = [ "random" ]
+
+[[test]]
+name = "test-debug"
+path = "tests/debug.rs"
+required-features = [ "debug" ]
+
+[[test]]
+name = "test-encoding"
+path = "tests/encoding.rs"
+required-features = [ "random" ]
+
+[[test]]
+name = "test-flow"
+path = "tests/flow.rs"
+required-features = [ "random" ]
+
+[[test]]
+name = "test-predicate"
+path = "tests/predicate.rs"
+required-features = [ "random" ]
