+++ src/interpreter/alu.rs
@@ -1,7 +1,7 @@
 use super::Interpreter;
 use crate::consts::*;
 
-use fuel_asm::{RegisterId, Word};
+use fuel_data::{RegisterId, Word};
 
 impl<S> Interpreter<S> {
     pub(crate) fn alu_overflow<B, C>(&mut self, ra: RegisterId, f: fn(B, C) -> (Word, bool), b: B, c: C) {
