+++ src/interpreter/blockchain.rs
@@ -1,9 +1,10 @@
-use super::{ExecuteError, Interpreter};
+use super::Interpreter;
 use crate::consts::*;
-use crate::data::{InterpreterStorage, MerkleStorage, Storage};
+use crate::data::InterpreterStorage;
+use crate::error::InterpreterError;
 
-use fuel_asm::{RegisterId, Word};
-use fuel_tx::{Address, Bytes32, Bytes8, Color, ContractId, Input, Salt};
+use fuel_data::{Address, Bytes32, Bytes8, Color, ContractId, RegisterId, Word};
+use fuel_tx::Input;
 
 use std::mem;
 
@@ -13,11 +14,11 @@ impl<S> Interpreter<S>
 where
     S: InterpreterStorage,
 {
-    pub(crate) fn coinbase(&self) -> Result<Address, ExecuteError> {
-        Ok(self.storage.coinbase()?)
+    pub(crate) fn coinbase(&self) -> Result<Address, InterpreterError> {
+        self.storage.coinbase().map_err(|e| e.into())
     }
 
-    pub(crate) fn burn(&mut self, a: Word) -> Result<(), ExecuteError> {
+    pub(crate) fn burn(&mut self, a: Word) -> Result<(), InterpreterError> {
         let (c, cx) = self.internal_contract_bounds()?;
 
         // Safety: Memory bounds logically verified by the interpreter
@@ -25,16 +26,17 @@ where
         let color = unsafe { Color::as_ref_unchecked(&self.memory[c..cx]) };
 
         let balance = self.balance(contract, color)?;
-        let balance = balance.checked_sub(a).ok_or(ExecuteError::NotEnoughBalance)?;
+        let balance = balance.checked_sub(a).ok_or(InterpreterError::NotEnoughBalance)?;
 
-        <S as MerkleStorage<ContractId, Color, Word>>::insert(&mut self.storage, contract, color, &balance)?;
+        self.storage
+            .merkle_contract_color_balance_insert(contract, color, balance)?;
 
         self.inc_pc();
 
         Ok(())
     }
 
-    pub(crate) fn mint(&mut self, a: Word) -> Result<(), ExecuteError> {
+    pub(crate) fn mint(&mut self, a: Word) -> Result<(), InterpreterError> {
         let (c, cx) = self.internal_contract_bounds()?;
 
         // Safety: Memory bounds logically verified by the interpreter
@@ -42,27 +44,28 @@ where
         let color = unsafe { Color::as_ref_unchecked(&self.memory[c..cx]) };
 
         let balance = self.balance(contract, color)?;
-        let balance = balance.checked_add(a).ok_or(ExecuteError::ArithmeticOverflow)?;
+        let balance = balance.checked_add(a).ok_or(InterpreterError::ArithmeticOverflow)?;
 
-        <S as MerkleStorage<ContractId, Color, Word>>::insert(&mut self.storage, contract, color, &balance)?;
+        self.storage
+            .merkle_contract_color_balance_insert(contract, color, balance)?;
 
         self.inc_pc();
 
         Ok(())
     }
 
-    pub(crate) fn code_copy(&mut self, a: Word, b: Word, c: Word, d: Word) -> Result<(), ExecuteError> {
+    pub(crate) fn code_copy(&mut self, a: Word, b: Word, c: Word, d: Word) -> Result<(), InterpreterError> {
         if d > MEM_MAX_ACCESS_SIZE
             || a > VM_MAX_RAM - d
-            || b > VM_MAX_RAM - ContractId::size_of() as Word
+            || b > VM_MAX_RAM - ContractId::LEN as Word
             || c > VM_MAX_RAM - d
         {
-            return Err(ExecuteError::MemoryOverflow);
+            return Err(InterpreterError::MemoryOverflow);
         }
 
         let (a, b, c, d) = (a as usize, b as usize, c as usize, d as usize);
 
-        let bx = b + ContractId::size_of();
+        let bx = b + ContractId::LEN;
         let cd = c + d;
 
         // Safety: Memory bounds are checked by the interpreter
@@ -74,10 +77,10 @@ where
             .iter()
             .any(|input| matches!(input, Input::Contract { contract_id, .. } if contract_id == contract))
         {
-            return Err(ExecuteError::ContractNotInTxInputs);
+            return Err(InterpreterError::ContractNotInTxInputs);
         }
 
-        let contract = self.contract(contract)?.ok_or(ExecuteError::ContractNotFound)?;
+        let contract = self.contract(contract)?.into_owned();
 
         if contract.as_ref().len() < d {
             self.try_zeroize(a, d)?;
@@ -90,8 +93,8 @@ where
         Ok(())
     }
 
-    pub(crate) fn block_hash(&mut self, a: Word, b: Word) -> Result<(), ExecuteError> {
-        let hash = self.storage.block_hash(b as u32)?;
+    pub(crate) fn block_hash(&mut self, a: Word, b: Word) -> Result<(), InterpreterError> {
+        let hash = self.storage.block_hash(b as u32).map_err(|e| e.into())?;
 
         self.try_mem_write(a as usize, hash.as_ref())?;
 
@@ -100,7 +103,7 @@ where
         Ok(())
     }
 
-    pub(crate) fn block_proposer(&mut self, a: Word) -> Result<(), ExecuteError> {
+    pub(crate) fn block_proposer(&mut self, a: Word) -> Result<(), InterpreterError> {
         self.coinbase()
             .and_then(|data| self.try_mem_write(a as usize, data.as_ref()))?;
 
@@ -109,19 +112,22 @@ where
         Ok(())
     }
 
-    pub(crate) fn code_root(&mut self, a: Word, b: Word) -> Result<(), ExecuteError> {
-        if a > VM_MAX_RAM - Bytes32::size_of() as Word || b > VM_MAX_RAM - ContractId::size_of() as Word {
-            return Err(ExecuteError::MemoryOverflow);
+    pub(crate) fn code_root(&mut self, a: Word, b: Word) -> Result<(), InterpreterError> {
+        if a > VM_MAX_RAM - Bytes32::LEN as Word || b > VM_MAX_RAM - ContractId::LEN as Word {
+            return Err(InterpreterError::MemoryOverflow);
         }
 
         let (a, b) = (a as usize, b as usize);
 
         // Safety: Memory bounds are checked by the interpreter
-        let contract_id = unsafe { ContractId::as_ref_unchecked(&self.memory[b..b + ContractId::size_of()]) };
+        let contract_id = unsafe { ContractId::as_ref_unchecked(&self.memory[b..b + ContractId::LEN]) };
 
-        let (_, root) = <S as Storage<ContractId, (Salt, Bytes32)>>::get(&self.storage, contract_id)
+        let (_, root) = self
+            .storage
+            .storage_contract_root(contract_id)
             .transpose()
-            .ok_or(ExecuteError::ContractNotFound)??;
+            .ok_or(InterpreterError::ContractNotFound)??
+            .into_owned();
 
         self.try_mem_write(a, root.as_ref())?;
 
@@ -130,29 +136,26 @@ where
         Ok(())
     }
 
-    pub(crate) fn code_size(&mut self, ra: RegisterId, b: Word) -> Result<(), ExecuteError> {
-        if b > VM_MAX_RAM - ContractId::size_of() as Word {
-            return Err(ExecuteError::MemoryOverflow);
+    pub(crate) fn code_size(&mut self, ra: RegisterId, b: Word) -> Result<(), InterpreterError> {
+        if b > VM_MAX_RAM - ContractId::LEN as Word {
+            return Err(InterpreterError::MemoryOverflow);
         }
 
         let b = b as usize;
 
         // Safety: Memory bounds are checked by the interpreter
-        let contract_id = unsafe { ContractId::as_ref_unchecked(&self.memory[b..b + ContractId::size_of()]) };
+        let contract_id = unsafe { ContractId::as_ref_unchecked(&self.memory[b..b + ContractId::LEN]) };
 
-        self.contract(contract_id)
-            .transpose()
-            .ok_or(ExecuteError::ContractNotFound)?
-            .map(|contract| self.registers[ra] = contract.as_ref().len() as Word)?;
+        self.registers[ra] = self.contract(contract_id)?.as_ref().as_ref().len() as Word;
 
         self.inc_pc();
 
         Ok(())
     }
 
-    pub(crate) fn state_read_word(&mut self, ra: RegisterId, b: Word) -> Result<(), ExecuteError> {
-        if b > VM_MAX_RAM - Bytes32::size_of() as Word {
-            return Err(ExecuteError::MemoryOverflow);
+    pub(crate) fn state_read_word(&mut self, ra: RegisterId, b: Word) -> Result<(), InterpreterError> {
+        if b > VM_MAX_RAM - Bytes32::LEN as Word {
+            return Err(InterpreterError::MemoryOverflow);
         }
 
         let b = b as usize;
@@ -160,10 +163,12 @@ where
         let contract = self.internal_contract()?;
 
         // Safety: Memory bounds are checked by the interpreter
-        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[b..b + Bytes32::size_of()]) };
+        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[b..b + Bytes32::LEN]) };
 
-        self.registers[ra] = <S as MerkleStorage<ContractId, Bytes32, Bytes32>>::get(&self.storage, &contract, key)?
-            .map(|state| unsafe { Bytes8::from_slice_unchecked(state.as_ref()).into() })
+        self.registers[ra] = self
+            .storage
+            .merkle_contract_state(&contract, key)?
+            .map(|state| unsafe { Bytes8::from_slice_unchecked(state.as_ref().as_ref()).into() })
             .map(Word::from_be_bytes)
             .unwrap_or(0);
 
@@ -172,9 +177,9 @@ where
         Ok(())
     }
 
-    pub(crate) fn state_read_qword(&mut self, a: Word, b: Word) -> Result<(), ExecuteError> {
-        if a > VM_MAX_RAM - Bytes32::size_of() as Word || b > VM_MAX_RAM - Bytes32::size_of() as Word {
-            return Err(ExecuteError::MemoryOverflow);
+    pub(crate) fn state_read_qword(&mut self, a: Word, b: Word) -> Result<(), InterpreterError> {
+        if a > VM_MAX_RAM - Bytes32::LEN as Word || b > VM_MAX_RAM - Bytes32::LEN as Word {
+            return Err(InterpreterError::MemoryOverflow);
         }
 
         let (a, b) = (a as usize, b as usize);
@@ -182,10 +187,13 @@ where
         let contract = self.internal_contract()?;
 
         // Safety: Memory bounds are checked by the interpreter
-        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[b..b + Bytes32::size_of()]) };
+        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[b..b + Bytes32::LEN]) };
 
-        let state =
-            <S as MerkleStorage<ContractId, Bytes32, Bytes32>>::get(&self.storage, &contract, key)?.unwrap_or_default();
+        let state = self
+            .storage
+            .merkle_contract_state(&contract, key)?
+            .map(|s| s.into_owned())
+            .unwrap_or_default();
 
         self.try_mem_write(a, state.as_ref())?;
 
@@ -194,9 +202,9 @@ where
         Ok(())
     }
 
-    pub(crate) fn state_write_word(&mut self, a: Word, b: Word) -> Result<(), ExecuteError> {
-        if a > VM_MAX_RAM - Bytes32::size_of() as Word {
-            return Err(ExecuteError::MemoryOverflow);
+    pub(crate) fn state_write_word(&mut self, a: Word, b: Word) -> Result<(), InterpreterError> {
+        if a > VM_MAX_RAM - Bytes32::LEN as Word {
+            return Err(InterpreterError::MemoryOverflow);
         }
 
         let a = a as usize;
@@ -204,22 +212,22 @@ where
 
         // Safety: Memory bounds logically verified by the interpreter
         let contract = unsafe { ContractId::as_ref_unchecked(&self.memory[c..cx]) };
-        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[a..a + Bytes32::size_of()]) };
+        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[a..a + Bytes32::LEN]) };
 
         let mut value = Bytes32::default();
 
         (&mut value[..WORD_SIZE]).copy_from_slice(&b.to_be_bytes());
 
-        <S as MerkleStorage<ContractId, Bytes32, Bytes32>>::insert(&mut self.storage, contract, key, &value)?;
+        self.storage.merkle_contract_state_insert(contract, key, &value)?;
 
         self.inc_pc();
 
         Ok(())
     }
 
-    pub(crate) fn state_write_qword(&mut self, a: Word, b: Word) -> Result<(), ExecuteError> {
-        if a > VM_MAX_RAM - Bytes32::size_of() as Word || b > VM_MAX_RAM - Bytes32::size_of() as Word {
-            return Err(ExecuteError::MemoryOverflow);
+    pub(crate) fn state_write_qword(&mut self, a: Word, b: Word) -> Result<(), InterpreterError> {
+        if a > VM_MAX_RAM - Bytes32::LEN as Word || b > VM_MAX_RAM - Bytes32::LEN as Word {
+            return Err(InterpreterError::MemoryOverflow);
         }
 
         let (a, b) = (a as usize, b as usize);
@@ -227,10 +235,10 @@ where
 
         // Safety: Memory bounds logically verified by the interpreter
         let contract = unsafe { ContractId::as_ref_unchecked(&self.memory[c..cx]) };
-        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[a..a + Bytes32::size_of()]) };
-        let value = unsafe { Bytes32::as_ref_unchecked(&self.memory[b..b + Bytes32::size_of()]) };
+        let key = unsafe { Bytes32::as_ref_unchecked(&self.memory[a..a + Bytes32::LEN]) };
+        let value = unsafe { Bytes32::as_ref_unchecked(&self.memory[b..b + Bytes32::LEN]) };
 
-        <S as MerkleStorage<ContractId, Bytes32, Bytes32>>::insert(&mut self.storage, &contract, key, value)?;
+        self.storage.merkle_contract_state_insert(contract, key, value)?;
 
         self.inc_pc();
