+++ src/interpreter/contract.rs
@@ -1,119 +1,36 @@
-use super::{ExecuteError, Interpreter};
-use crate::crypto;
-use crate::data::{InterpreterStorage, MerkleStorage, Storage};
+use super::Interpreter;
+use crate::contract::Contract;
+use crate::data::InterpreterStorage;
+use crate::error::InterpreterError;
 
-use fuel_asm::Word;
-use fuel_tx::crypto::Hasher;
-use fuel_tx::{Bytes32, Color, ContractId, Salt, Transaction, ValidationError};
+use fuel_data::{Color, ContractId, Word};
 
-use std::cmp;
-use std::convert::TryFrom;
-
-#[derive(Debug, Default, Clone, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serde-types", derive(serde::Serialize, serde::Deserialize))]
-pub struct Contract(Vec<u8>);
-
-impl Contract {
-    pub fn root(&self) -> Bytes32 {
-        let root = self.0.chunks(8).map(|c| {
-            let mut bytes = [0u8; 8];
-
-            let l = cmp::min(c.len(), 8);
-            (&mut bytes[..l]).copy_from_slice(c);
-
-            bytes
-        });
-
-        crypto::ephemeral_merkle_root(root)
-    }
-
-    pub fn id(&self, salt: &Salt, root: &Bytes32) -> ContractId {
-        let mut hasher = Hasher::default();
-
-        hasher.input(ContractId::SEED);
-        hasher.input(salt);
-        hasher.input(root);
-
-        ContractId::from(*hasher.digest())
-    }
-}
-
-impl From<Vec<u8>> for Contract {
-    fn from(c: Vec<u8>) -> Self {
-        Self(c)
-    }
-}
-
-impl From<&[u8]> for Contract {
-    fn from(c: &[u8]) -> Self {
-        Self(c.into())
-    }
-}
-
-impl From<&mut [u8]> for Contract {
-    fn from(c: &mut [u8]) -> Self {
-        Self(c.into())
-    }
-}
-
-impl From<Contract> for Vec<u8> {
-    fn from(c: Contract) -> Vec<u8> {
-        c.0
-    }
-}
-
-impl AsRef<[u8]> for Contract {
-    fn as_ref(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl AsMut<[u8]> for Contract {
-    fn as_mut(&mut self) -> &mut [u8] {
-        self.0.as_mut()
-    }
-}
-
-impl TryFrom<&Transaction> for Contract {
-    type Error = ExecuteError;
-
-    fn try_from(tx: &Transaction) -> Result<Self, Self::Error> {
-        match tx {
-            Transaction::Create {
-                bytecode_witness_index,
-                witnesses,
-                ..
-            } => witnesses
-                .get(*bytecode_witness_index as usize)
-                .map(|c| c.as_ref().into())
-                .ok_or(ValidationError::TransactionCreateBytecodeWitnessIndex.into()),
-
-            _ => Err(ValidationError::TransactionScriptOutputContractCreated { index: 0 }.into()),
-        }
-    }
-}
+use std::borrow::Cow;
 
 impl<S> Interpreter<S>
 where
     S: InterpreterStorage,
 {
-    pub(crate) fn contract(&self, contract: &ContractId) -> Result<Option<Contract>, ExecuteError> {
-        Ok(<S as Storage<ContractId, Contract>>::get(&self.storage, contract)?)
+    pub(crate) fn contract(&self, contract: &ContractId) -> Result<Cow<'_, Contract>, InterpreterError> {
+        self.storage
+            .storage_contract(contract)
+            .transpose()
+            .ok_or(InterpreterError::ContractNotFound)?
     }
 
-    pub(crate) fn check_contract_exists(&self, contract: &ContractId) -> Result<bool, ExecuteError> {
-        Ok(<S as Storage<ContractId, Contract>>::contains_key(
-            &self.storage,
-            contract,
-        )?)
+    pub(crate) fn check_contract_exists(&self, contract: &ContractId) -> Result<bool, InterpreterError> {
+        self.storage.storage_contract_exists(contract)
     }
 
-    pub(crate) fn balance(&self, contract: &ContractId, color: &Color) -> Result<Word, ExecuteError> {
-        Ok(<S as MerkleStorage<ContractId, Color, Word>>::get(&self.storage, contract, color)?.unwrap_or(0))
+    pub(crate) fn balance(&self, contract: &ContractId, color: &Color) -> Result<Word, InterpreterError> {
+        Ok(self
+            .storage
+            .merkle_contract_color_balance(contract, color)?
+            .unwrap_or_default())
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "random"))]
 mod tests {
     use crate::consts::*;
     use crate::prelude::*;
@@ -126,8 +43,7 @@ mod tests {
 
         let mut balance = 1000;
 
-        let storage = MemoryStorage::default();
-        let mut vm = Interpreter::with_storage(storage);
+        let mut vm = Interpreter::in_memory();
 
         let gas_price = 0;
         let gas_limit = 1_000_000;
