+++ src/interpreter/executors.rs
@@ -1,11 +1,6 @@
-mod balances;
-mod initialization;
 mod instruction;
 mod main;
 mod predicate;
-mod state;
 
 #[cfg(feature = "debug")]
 mod debug;
-
-pub use state::{ExecuteState, ProgramState, StateTransition, StateTransitionRef};
