+++ src/interpreter/gas.rs
@@ -1,53 +1,10 @@
-use super::{ExecuteError, Interpreter};
+use super::Interpreter;
 use crate::consts::*;
+use crate::error::InterpreterError;
+use crate::gas::GasUnit;
 
-use fuel_asm::{Opcode, Word};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum GasUnit {
-    Atom(Word),
-    Arithmetic(Word),
-    ArithmeticExpensive(Word),
-    RegisterRead(Word),
-    RegisterWrite(Word),
-    Branching(Word),
-    MemoryOwnership(Word),
-    MemoryWrite(Word),
-    Accumulated(Word),
-    Undefined,
-}
-
-impl GasUnit {
-    pub const fn cost(&self) -> Word {
-        use GasUnit::*;
-
-        match self {
-            Atom(1) => 1,
-            Arithmetic(1) => 5,
-            ArithmeticExpensive(1) => 7,
-            RegisterRead(1) => 1,
-            RegisterWrite(1) => 2,
-            Branching(1) => 10,
-            MemoryOwnership(1) => 9,
-            MemoryWrite(1) => 8,
-            Undefined => 20,
-
-            Atom(n) => *n * Atom(1).cost(),
-            Arithmetic(n) => *n * Arithmetic(1).cost(),
-            ArithmeticExpensive(n) => *n * ArithmeticExpensive(1).cost(),
-            RegisterRead(n) => *n * RegisterRead(1).cost(),
-            RegisterWrite(n) => *n * RegisterWrite(1).cost(),
-            Branching(n) => *n * Branching(1).cost(),
-            MemoryOwnership(n) => *n * MemoryOwnership(1).cost(),
-            MemoryWrite(n) => *n * MemoryWrite(1).cost(),
-            Accumulated(c) => *c,
-        }
-    }
-
-    pub const fn join(self, other: Self) -> Self {
-        Self::Accumulated(self.cost() + other.cost())
-    }
-}
+use fuel_asm::Opcode;
+use fuel_data::Word;
 
 impl<S> Interpreter<S> {
     /// Calculate the gas cost for the current runtime state.
@@ -121,14 +78,14 @@ impl<S> Interpreter<S> {
 
     // TODO enable const flag
     // https://github.com/rust-lang/rust/issues/57349
-    pub(crate) fn gas_charge(&mut self, op: &Opcode) -> Result<(), ExecuteError> {
+    pub(crate) fn gas_charge(&mut self, op: &Opcode) -> Result<(), InterpreterError> {
         let cost = !self.is_predicate() as Word * self.gas_cost(op);
 
         if cost > self.registers[REG_CGAS] {
             self.registers[REG_GGAS] -= self.registers[REG_CGAS];
             self.registers[REG_CGAS] = 0;
 
-            Err(ExecuteError::OutOfGas)
+            Err(InterpreterError::OutOfGas)
         } else {
             self.registers[REG_CGAS] -= cost;
