--- tests/encoding/interpreter.rs
@@ -1,38 +1,0 @@
-use super::assert_encoding_correct;
-use fuel_vm::consts::*;
-use fuel_vm::prelude::*;
-use rand::rngs::StdRng;
-use rand::{Rng, SeedableRng};
-
-#[test]
-fn call() {
-    let rng = &mut StdRng::seed_from_u64(2322u64);
-
-    assert_encoding_correct(
-        (0..10)
-            .map(|_| Call::new(rng.gen(), rng.gen(), rng.gen()))
-            .collect::<Vec<Call>>()
-            .as_slice(),
-    );
-}
-
-#[test]
-fn call_frame() {
-    let rng = &mut StdRng::seed_from_u64(2322u64);
-
-    assert_encoding_correct(
-        (0..10)
-            .map(|_| {
-                CallFrame::new(
-                    rng.gen(),
-                    rng.gen(),
-                    [rng.gen(); VM_REGISTER_COUNT],
-                    rng.gen(),
-                    rng.gen(),
-                    vec![rng.gen(); 200].into(),
-                )
-            })
-            .collect::<Vec<CallFrame>>()
-            .as_slice(),
-    );
-}
