--- tests/interpreter/mod.rs
@@ -1,8 +1,0 @@
-mod alu;
-mod blockchain;
-mod crypto;
-mod flow;
-mod predicate;
-
-#[cfg(feature = "debug")]
-mod debug;
