--- tests/vm.rs
@@ -1,2 +1,0 @@
-mod encoding;
-mod interpreter;
